feat: remove wildcard return type on Controller
feat: move all dummy dto creator method to MockFactoryUtil.java
