+++ api/src/main/java/com/bbolab/gaonna/api/v1/controller/ArticleController.java
@@ -3,7 +3,6 @@ package com.bbolab.gaonna.api.v1.controller;
 import com.bbolab.gaonna.api.v1.dto.article.ArticleCreateRequestDto;
 import com.bbolab.gaonna.api.v1.dto.article.ArticleResponseDto;
 import com.bbolab.gaonna.api.v1.dto.article.ArticleUpdateRequestDto;
-import com.bbolab.gaonna.api.v1.dto.comment.CommentResponseDto;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
@@ -22,9 +21,9 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import java.net.URI;
-import java.time.LocalDateTime;
-import java.util.Collections;
 import java.util.UUID;
+
+import static com.bbolab.gaonna.api.v1.controller.MockFactoryUtil.createDummyArticleResponseDto;
 
 @Api(value = "article")
 @RestController
@@ -37,7 +36,7 @@ public class ArticleController {
     @ApiOperation(value = "Searching article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success", response = ArticleResponseDto.class)})
     @GetMapping("{articleId}")
-    public ResponseEntity<?> get(@ApiParam(value = "ex) 72f92a8b-1866-4f08-bdf1-5c4826d0378b", required = true) @PathVariable String articleId) {
+    public ResponseEntity<ArticleResponseDto> get(@ApiParam(value = "ex) 72f92a8b-1866-4f08-bdf1-5c4826d0378b", required = true) @PathVariable String articleId) {
         ArticleResponseDto dto = createDummyArticleResponseDto();
         dto.setArticleId(articleId);
         return ResponseEntity.ok(dto);
@@ -46,7 +45,7 @@ public class ArticleController {
     @ApiOperation(value = "Create new article")
     @ApiResponses({@ApiResponse(code = 201, message = "Success", response = ArticleResponseDto.class)})
     @PostMapping
-    public ResponseEntity<?> create(@RequestBody ArticleCreateRequestDto requestDto) {
+    public ResponseEntity<ArticleResponseDto> create(@RequestBody ArticleCreateRequestDto requestDto) {
         ArticleResponseDto dto = createDummyArticleResponseDto();
         modelMapper.map(requestDto, dto);
         dto.setArticleId(UUID.randomUUID().toString());
@@ -56,7 +55,7 @@ public class ArticleController {
     @ApiOperation(value = "Update article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success", response = ArticleResponseDto.class)})
     @PutMapping("{articleId}")
-    public ResponseEntity<?> update(@PathVariable String articleId, @RequestBody ArticleUpdateRequestDto requestDto) {
+    public ResponseEntity<ArticleResponseDto> update(@PathVariable String articleId, @RequestBody ArticleUpdateRequestDto requestDto) {
         // TODO : need article owner checking
         ArticleResponseDto dto = createDummyArticleResponseDto();
         modelMapper.map(requestDto, dto);
@@ -70,14 +69,14 @@ public class ArticleController {
     @ApiOperation(value = "Delete article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success")})
     @DeleteMapping("{articleId}")
-    public ResponseEntity<?> delete(@PathVariable String articleId){
+    public ResponseEntity<Void> delete(@PathVariable String articleId){
         return ResponseEntity.ok().build();
     }
 
     @ApiOperation(value = "User add like to article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success")})
     @PostMapping("/like/{articleId}")
-    public ResponseEntity<?> addLike(@PathVariable String articleId) {
+    public ResponseEntity<Void> addLike(@PathVariable String articleId) {
         // TODO : 해당 요청을 한 유저가 Article에 좋아요를 이미 눌렀을 경우는 예외
         return ResponseEntity.ok().build();
     }
@@ -85,24 +84,8 @@ public class ArticleController {
     @ApiOperation(value = "User cancel like to article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success")})
     @DeleteMapping("/like/{articleId}")
-    public ResponseEntity<?> deleteLike(@PathVariable String articleId) {
+    public ResponseEntity<Void> deleteLike(@PathVariable String articleId) {
         // TODO : 해당 요청을 한 유저가 Article에 좋아요를 누르지 않았을 경우는 예외
         return ResponseEntity.ok().build();
     }
-
-    public static ArticleResponseDto createDummyArticleResponseDto() {
-        return ArticleResponseDto.builder()
-                .articleId(UUID.randomUUID().toString())
-                .title("test-title")
-                .content("test-content-with-html-format")
-                .updatedTime(LocalDateTime.now())
-                .likeCount(3)
-                .commentCount(1)
-                .comments(Collections.singletonList(CommentResponseDto.builder()
-                        .memberId(UUID.randomUUID().toString())
-                        .memberName("test-member-name")
-                        .content("test-comment-content")
-                        .updatedTime(LocalDateTime.now()).build())).build();
-    }
-
 }
