+++ api/src/main/java/com/bbolab/gaonna/api/v1/controller/CommentController.java
@@ -19,10 +19,10 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
-
-import java.time.LocalDateTime;
 import java.util.Collections;
-import java.util.UUID;
+
+import static com.bbolab.gaonna.api.v1.controller.MockFactoryUtil.createDummyComment;
+import static com.bbolab.gaonna.api.v1.controller.MockFactoryUtil.createDummyCommentResponseDto;
 
 @Api(value = "comment")
 @Controller
@@ -35,18 +35,19 @@ public class CommentController {
     @ApiOperation(value = "Searching article's comments")
     @ApiResponses({@ApiResponse(code = 200, message = "Success", response = CommentListResponseDto.class)})
     @GetMapping
-    public ResponseEntity<?> getAll(@ApiParam(value = "ex) 72f92a8b-1866-4f08-bdf1-5c4826d0378b", required = true) @PathVariable String articleId) {
+    public ResponseEntity<CommentListResponseDto> getAll(@ApiParam(value = "ex) 72f92a8b-1866-4f08-bdf1-5c4826d0378b", required = true) @PathVariable String articleId) {
         CommentResponseDto comment = createDummyComment();
-        CommentListResponseDto dto = CommentListResponseDto.builder().build();
-        dto.setCommentLists(Collections.singletonList(comment));
-        dto.setArticleId(articleId);
+        CommentListResponseDto dto = CommentListResponseDto.builder()
+                .commentLists(Collections.singletonList(comment))
+                .articleId(articleId)
+                .build();
         return ResponseEntity.ok().body(dto);
     }
 
     @ApiOperation(value = "Update comment on article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success", response = CommentResponseDto.class)})
     @PutMapping("{commentId}")
-    public ResponseEntity<?>  updateComment(@PathVariable String articleId, @PathVariable String commentId, @RequestBody CommentCreateUpdateRequestDto requestDto) {
+    public ResponseEntity<CommentResponseDto>  updateComment(@PathVariable String articleId, @PathVariable String commentId, @RequestBody CommentCreateUpdateRequestDto requestDto) {
         CommentResponseDto dto = createDummyComment();
         modelMapper.map(requestDto, dto);
         dto.setCommentId(commentId);
@@ -56,7 +57,7 @@ public class CommentController {
     @ApiOperation(value = "Create comment on article")
     @ApiResponses({@ApiResponse(code = 200, message = "Success", response = CommentResponseDto.class)})
     @PostMapping
-    public ResponseEntity<?> createComment(@PathVariable String articleId, @RequestBody CommentCreateUpdateRequestDto requestDto) {
+    public ResponseEntity<CommentResponseDto> createComment(@PathVariable String articleId, @RequestBody CommentCreateUpdateRequestDto requestDto) {
         CommentResponseDto dto = createDummyCommentResponseDto();
         modelMapper.map(requestDto, dto);
         return ResponseEntity.ok().body(dto);
@@ -65,24 +66,7 @@ public class CommentController {
     @ApiOperation(value = "Delete comment on article")
     @DeleteMapping("{commentId}")
     @ApiResponses({@ApiResponse(code = 200, message = "Success")})
-    public ResponseEntity<?> deleteComment(@PathVariable String articleId, @PathVariable String commentId) {
+    public ResponseEntity<Void> deleteComment(@PathVariable String articleId, @PathVariable String commentId) {
         return ResponseEntity.ok().build();
-    }
-
-    public static CommentResponseDto createDummyComment() {
-        return CommentResponseDto.builder()
-                .memberId(UUID.randomUUID().toString())
-                .memberName("test-name")
-                .content("test-comment-content")
-                .updatedTime(LocalDateTime.now()).build();
-    }
-
-    public static CommentResponseDto createDummyCommentResponseDto() {
-        return CommentResponseDto.builder()
-                .commentId(UUID.randomUUID().toString())
-                .memberId(UUID.randomUUID().toString())
-                .memberName("test-member-name")
-                .content("test-comment-content")
-                .updatedTime(LocalDateTime.now()).build();
     }
 }
